dom/tests/mochitest/browser-frame/test_browserFrame3.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=710231
-->
<head>
  <title>Test for Bug 710231</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserFrameHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=710231">Mozilla Bug 710231</a>

<!--
  Test that an iframe with the |mozbrowser| attribute does not emit
  mozbrowserX events when this page is not in the whitelist.
-->

<script type="application/javascript;version=1.7">
"use strict";

SimpleTest.waitForExplicitFinish();

function runTest() {
  browserFrameHelpers.setEnabledPref(true);
  browserFrameHelpers.setWhitelistPref(' http://foobar.com');

  var iframe = document.createElement('iframe');
  iframe.mozbrowser = true;
  document.body.appendChild(iframe);

  iframe.addEventListener('mozbrowserloadstart', function() {
    ok(false, 'Should not send mozbrowserloadstart event.');
  });

  iframe.addEventListener('load', function() {
    ok(true, 'Got iframe load event.');
    SimpleTest.finish();
  });

  iframe.src = browserFrameHelpers.emptyPage1;
}

addEventListener('load', function() { SimpleTest.executeSoon(runTest); });

</script>

</body>
</html>
